Bug Fixes/ Rename Method
I renamed the doneButton method to toggle so that it makes more sense
I fixed a bug if you tried to toggle +/- with a blank text field it
would crash. Now it does nothing.
I made sure that if the textfield is blank when editing you have to be
the delegate in order to change the text to $0.00